Move storage module higher up
