+++ openverse_catalog/dags/provider_api_scripts/brooklyn_museum.py
@@ -4,7 +4,7 @@ import os
 import lxml.html as html
 from common.licenses.licenses import get_license_info
 from common.requester import DelayedRequester
-from common.storage.image import ImageStore
+from storage.image import ImageStore
 from util.loader import provider_details as prov
