+++ openverse_catalog/dags/provider_api_scripts/raw_pixel.py
@@ -4,7 +4,7 @@ from urllib.parse import parse_qs, urlparse
 import requests
 from common.licenses.licenses import get_license_info
 from common.requester import DelayedRequester
-from common.storage.image import ImageStore
+from storage.image import ImageStore
 from util.loader import provider_details as prov
