+++ openverse_catalog/dags/provider_api_scripts/stocksnap.py
@@ -15,7 +15,7 @@ import logging
 
 from common.licenses.licenses import get_license_info
 from common.requester import DelayedRequester
-from common.storage.image import ImageStore
+from storage.image import ImageStore
 from util.loader import provider_details as prov
