+++ openverse_catalog/dags/provider_api_scripts/wikimedia_commons.py
@@ -20,7 +20,7 @@ from urllib.parse import urlparse
 import lxml.html as html
 from common.licenses.licenses import get_license_info
 from common.requester import DelayedRequester
-from common.storage.image import ImageStore
+from storage.image import ImageStore
 from util.loader import provider_details as prov
